added the preferences setup screen

+++ gui/public/electron.js
@@ -13,6 +13,8 @@ const production = process.env.ELECTRON_START_URL === undefined
 
 const Store = require('electron-store');
 const store = new Store();
+let port = 8081;
+
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
@@ -20,18 +22,31 @@ let mainWindow;
 // Do the same for the backend web server.
 let backendServer
 
+
+function urlPath(p) {
+  var editUrl;
+  if (process.env.ELECTRON_START_URL) 
+    editUrl = path.join(process.env.ELECTRON_START_URL , "#", p)
+  else
+    editUrl = url.format({
+      pathname: path.join(__dirname, '/../build/index.html'),
+      hash: p,
+      protocol: 'file:',
+      slashes: true
+    });
+  return editUrl;
+};
+
 function createWindow() {
     // Create the browser window.
     mainWindow = new BrowserWindow({titleBarStyle: 'hidden-inset', width: 800, height: 600, show: false});
+    mainWindow.port = port
 
     // and load the index.html of the app.
 
-    const startUrl = process.env.ELECTRON_START_URL || url.format({
-        pathname: path.join(__dirname, '/../build/index.html'),
-        protocol: 'file:',
-        slashes: true
-    });
+    const startUrl = urlPath('/');
     mainWindow.loadURL(startUrl);
+
 
     // Open the DevTools.
     // if(production) 
@@ -63,16 +78,37 @@ function createBackendServer () {
     path.join(appRootDir, 'resources');
 
   const cmd = path.join(execPath, 'calculus-toolbox');
-  // backendServer = child_process.spawn(cmd, ['gui', '--path', '../calculi', '--name' , 'Sequent']);
-  backendServer = child_process.spawn(cmd, ['gui', '--path', '../calculi']);
+  var workDir = store.get('workDir')
 
-  backendServer.stdout.on('data', (data) => {
-    console.log(`stdout: ${data}`);
-  });
+  const launch = () => {
+    const currentCalc = store.get('currentCalc')
+    if(currentCalc && currentCalc !== '') backendServer = child_process.spawn(cmd, ['gui', '--path', workDir, '--calc', currentCalc]);
+    else backendServer = child_process.spawn(cmd, ['gui', '--path', workDir]);
 
-  backendServer.stderr.on('data', (data) => {
-    console.log(`stderr: ${data}`);
-  });
+    backendServer.stdout.on('data', (data) => {
+      console.log(`stdout: ${data}`);
+    });
+
+    backendServer.stderr.on('data', (data) => {
+      console.log(`stderr: ${data}`);
+    });
+    createWindow();
+  }
+
+  console.log(workDir)
+  if(workDir === undefined || workDir === ''){
+    var setupWindow = new BrowserWindow({ width: 800, height: 300, title:"Modify calculus" });
+
+    setupWindow.loadURL(urlPath('/preferences/initial'));
+    setupWindow.setMenu(null);
+
+    setupWindow.on('close', function () {
+      setupWindow = null
+      workDir = store.get('workDir')
+      if(workDir && workDir !== '') launch()
+      else app.quit();
+    })      
+  } else launch() 
 }
 
 // Start the backend web server when Electron has finished initializing.
@@ -80,10 +116,10 @@ app.on('ready', createBackendServer)
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
-app.on('ready', createWindow);
+// app.on('ready', createWindow);
 
 app.on('will-quit', function() {
-  backendServer.kill()
+  if(backendServer) backendServer.kill()
 })
 
 // Quit when all windows are closed.
@@ -99,7 +135,7 @@ app.on('activate', function () {
     // On OS X it's common to re-create a window in the app when the
     // dock icon is clicked and there are no other windows open.
     if (mainWindow === null) {
-        createWindow()
+        createWindow(8081)
     }
 });
